// src/types/technologies.tsx
export enum Technology {
  CSS = 'CSS',
  Express = 'Express',
  Firebase = 'Firebase',
  HTML5 = 'HTML',
  JavaScript = 'JavaScript',
  Node = 'Node.js',
  React = 'React',
  ReactRouter = 'React Router',
  TypeScript = 'TypeScript',
  Jest = 'Jest',
  Webpack = 'Webpack',
  MaterialUi = 'Material UI',
}